refactor(landing): promote home2 components to site/
These are the site's design language, not homepage-local. Phase 2-4 pages
consume them, so the name was actively misleading. Move only — no behaviour
change.
