+++ apps/landing/src/components/site/CommunityLoop.tsx
@@ -2,7 +2,7 @@ import type { ReactNode } from 'react'
 import { motion } from 'framer-motion'
 import { Link } from 'react-router-dom'
 import { Mascot } from '@/components/ui/mascot'
-import { HomeSection, SectionTitle } from '@/components/sections/home2/primitives'
+import { HomeSection, SectionTitle } from '@/components/site/primitives'
 import { trackLandingEvent } from '@/lib/analytics'
 
 const EASE = [0.16, 1, 0.3, 1] as const
