+++ apps/landing/src/components/site/PricingTeaser.tsx
@@ -1,7 +1,7 @@
 import { motion } from 'framer-motion'
 import { Link } from 'react-router-dom'
 import { ArrowRight, Check } from 'lucide-react'
-import { HomeSection, SectionTitle } from '@/components/sections/home2/primitives'
+import { HomeSection, SectionTitle } from '@/components/site/primitives'
 import { SYNC_PLAN_TIERS, type SyncPlanId, type SyncPlanTier } from '@/lib/constants'
 import { trackLandingEvent } from '@/lib/analytics'
 import { cn } from '@/lib/utils'
