+++ apps/landing/src/components/site/widgets/NoteEditorWidget.tsx
